pyproject.toml
[project]
name = "podalirius"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "celery==5.5.3",
  "redis==6.4.0",
  "fastapi==0.116.1",
  "pytest==8.4.1",
  "httpie==3.2.4",
  "uvicorn==0.35.0",
  "python-jose==3.5.0",
  "passlib==1.7.4",
  "python-dotenv==1.1.1",
  "pytest-dotenv==0.5.2",
  "sqlmodel==0.0.24",
  "PyMySQL==1.1.2",
  "cryptography==45.0.6",
  "pydantic-settings==2.10.1",
  "httpx==0.28.1",
  "fastapi-utils==0.8.0",
  "typing-inspect==0.9.0",
]

[build-system]
requires = ["setuptools>=61"]
bduild-backend = "setuptools.build_meta"
